Move golinkcontext pkg into backend
