+++ backend/golink_service.go
@@ -9,9 +9,9 @@ import (
 	"cloud.google.com/go/firestore"
 	"github.com/bufbuild/connect-go"
 	golinkv1 "github.com/nownabe/golink/backend/gen/golink/v1"
+	"github.com/nownabe/golink/backend/golinkcontext"
 	"github.com/nownabe/golink/go/clog"
 	"github.com/nownabe/golink/go/errors"
-	"github.com/nownabe/golink/go/golinkcontext"
 	"golang.org/x/exp/slices"
 )
